refactor: split game_manager attack tests by hero
Move Elara/Thraïn/Thalia/Azrak-specific attack regression tests out
of attack_flow.rs's test module into per-hero sibling files, leaving
only hero-agnostic combat-mechanics tests in attack_flow.rs itself.
No behavior change — same 348 passing tests.
