+++ src/server/game_manager/attack_flow.rs
@@ -407,10 +407,22 @@ impl GameManager {
 }
 
 #[cfg(test)]
+#[path = "azrak_tests.rs"]
+mod azrak_tests;
+#[cfg(test)]
+#[path = "elara_tests.rs"]
+mod elara_tests;
+#[cfg(test)]
+#[path = "thalia_tests.rs"]
+mod thalia_tests;
+#[cfg(test)]
+#[path = "thrain_tests.rs"]
+mod thrain_tests;
+
+#[cfg(test)]
 mod tests {
     use crate::character_mod::attack_type::AttackType;
     use crate::character_mod::buffers::{BufKinds, Buffer};
-
     use crate::character_mod::class::Class;
     use crate::character_mod::rank::Rank;
     use crate::common::constants::attak_const::COEFF_CRIT_DMG;
@@ -420,6 +432,7 @@ mod tests {
     use crate::server::game_state::GameStatus;
     use crate::testing::testing_all_characters::{self, testing_test_ally1_vs_test_boss1};
     use crate::{common::constants::stats_const::*, testing::testing_atk::*};
+
     #[test]
     fn unit_launch_attack_none_atk_hero() {
         let (mut gm, _hero_launcher_id_name, _target_id_name) = testing_test_ally1_vs_test_boss1();
@@ -899,86 +912,6 @@ mod tests {
     }
 
     #[test]
-    fn unit_eclat_despoir_buffs_thrain_physical_and_magical_power() {
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance using new_round() (no attacks fired) so Thraïn's stats stay at
-        // their equipment-only baseline with no accumulated combat buffs.
-        let mut max_setup = 30;
-        while !gm.pm.current_player.id_name.contains("Elara") && max_setup > 0 {
-            gm.new_round();
-            max_setup -= 1;
-        }
-        if !gm.pm.current_player.id_name.contains("Elara") {
-            return;
-        }
-
-        // Disable crit and ensure full mana for determinism.
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        gm.pm.current_player.stats.all_stats[MANA].current = mana_max;
-
-        let thrain_id = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name.contains("Thraïn"))
-            .map(|h| h.id_name.clone())
-            .expect("Thraïn must be among the lotr heroes");
-
-        let old_phy_pow = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[PHYSICAL_POWER]
-            .max;
-        let old_mag_pow = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_POWER]
-            .max;
-
-        // Thraïn's magical power comes from equipment (max_raw == 0 but equip buffers
-        // give a non-zero effective max).  Both stats must be non-zero to make this
-        // test meaningful.
-        assert!(
-            old_mag_pow > 0,
-            "Thraïn must have non-zero magical power from equipment"
-        );
-        assert!(old_phy_pow > 0, "Thraïn must have non-zero physical power");
-
-        gm.launch_attack(Some("Eclat d'espoir"));
-
-        // Both stats must be boosted by +15 % (integer arithmetic matches the engine).
-        assert_eq!(
-            old_phy_pow + (0.15 * old_phy_pow as f64) as u64,
-            gm.pm
-                .get_active_hero_character(&thrain_id)
-                .unwrap()
-                .stats
-                .all_stats[PHYSICAL_POWER]
-                .max,
-            "Eclat d'espoir should boost Thraïn physical power by 15 %"
-        );
-        assert_eq!(
-            old_mag_pow + (0.15 * old_mag_pow as f64) as u64,
-            gm.pm
-                .get_active_hero_character(&thrain_id)
-                .unwrap()
-                .stats
-                .all_stats[MAGICAL_POWER]
-                .max,
-            "Eclat d'espoir should boost Thraïn magical power by 15 %"
-        );
-    }
-
-    #[test]
     fn unit_launch_attack_end_of_effect() {
         let (mut gm, hero_launcher_id_name, _target_id_name) = testing_test_ally1_vs_test_boss1();
 
@@ -1380,89 +1313,6 @@ mod tests {
         }
     }
 
-    /// Fracas Marteau deals self-damage via its buffer effects on the caster.
-    /// This test verifies that a low-HP hero is killed by the self-damage component.
-    #[test]
-    fn unit_fracas_marteau_can_kill_caster() {
-        use crate::{
-            character_mod::{attack_type::AttackType, buffers::BufKinds, effect::EffectParam},
-            common::constants::{
-                all_target_const::TARGET_HIMSELF, reach_const::INDIVIDUAL, stats_const::HP,
-            },
-        };
-
-        let (mut gm, hero_id_name, _) = testing_test_ally1_vs_test_boss1();
-
-        // Build a self-damage attack: 50 HP self-damage (guaranteed kill at 10 HP)
-        use crate::character_mod::buffers::Buffer;
-        let fracas_marteau = AttackType {
-            name: "Fracas Marteau".to_owned(),
-            target: TARGET_HIMSELF.to_owned(),
-            reach: INDIVIDUAL.to_owned(),
-            all_effects: vec![EffectParam {
-                nb_turns: 1,
-                target_kind: TARGET_HIMSELF.to_owned(),
-                reach: INDIVIDUAL.to_owned(),
-                buffer: Buffer {
-                    kind: BufKinds::ChangeCurrentStat,
-                    value: -50,
-                    is_percent: false,
-                    stats_name: HP.to_owned(),
-                    is_passive_enabled: false,
-                    is_passive: false,
-                },
-                ..Default::default()
-            }],
-            ..Default::default()
-        };
-
-        // Set hero HP to 10 so self-damage is lethal
-        for hero in gm.pm.active_heroes.iter_mut() {
-            if hero.id_name == hero_id_name {
-                hero.stats.get_mut_value(HP).current = 10;
-                hero.attacks_list
-                    .insert(fracas_marteau.name.clone(), fracas_marteau.clone());
-            }
-        }
-        // Also update current_player (shadow copy)
-        if gm.pm.current_player.id_name == hero_id_name {
-            gm.pm.current_player.stats.get_mut_value(HP).current = 10;
-            gm.pm
-                .current_player
-                .attacks_list
-                .insert(fracas_marteau.name.clone(), fracas_marteau.clone());
-        }
-
-        let result = gm.launch_attack(Some(&fracas_marteau.name));
-
-        // The attack must have been launched by our hero
-        assert_eq!(
-            result.launcher_id_name, hero_id_name,
-            "Fracas Marteau should be launched by {hero_id_name}"
-        );
-        // There must be at least one HP effect on the caster
-        assert!(
-            !result.new_game_atk_effects.is_empty(),
-            "Fracas Marteau should produce at least one game effect"
-        );
-
-        // The hero should be dead after taking 50+ self-damage from 10 HP
-        let hero_after = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name == hero_id_name);
-        if let Some(hero) = hero_after {
-            assert!(
-                hero.stats.is_dead() == Some(true) || hero.stats.all_stats[HP].current == 0,
-                "Fracas Marteau should kill the hero at 10 HP, but HP is {}",
-                hero.stats.all_stats[HP].current
-            );
-        }
-    }
-
-    // ── Aggro integration tests ────────────────────────────────────────────────
-
     /// After a damage attack the launcher's aggro should be strictly greater than its initial value.
     #[test]
     fn unit_aggro_increases_after_damage_attack() {
@@ -1698,2021 +1548,6 @@ mod tests {
         assert!(
             aggro_after_turn2 > 0,
             "Aggro should be positive after two attacks: turn2={aggro_after_turn2}"
-        );
-    }
-
-    /// Aggro accumulates correctly for a real LOTR hero (Thraïn) using "Frappe Cinglante"
-    /// across two consecutive turns.  Uses dxrpg_game_manager() so actual hero data is tested.
-    #[test]
-    fn unit_aggro_thrain_frappe_cinglante_accumulates() {
-        use crate::common::constants::stats_const::AGGRO;
-        use crate::server::game_state::GameStatus;
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance until Thraïn is the current player.
-        let mut max_setup = 30;
-        while !gm.pm.current_player.id_name.contains("Thraïn")
-            && gm.game_state.status != GameStatus::EndOfGame
-            && gm.game_state.status != GameStatus::EndOfScenario
-            && max_setup > 0
-        {
-            gm.launch_attack(None);
-            max_setup -= 1;
-        }
-        // Skip if Thraïn isn't up, or if the scenario already ended on Linux (bosses
-        // died before Thraïn's first turn — no valid targets remain for the assertion).
-        if !gm.pm.current_player.id_name.contains("Thraïn")
-            || gm.game_state.status != GameStatus::StartRound
-        {
-            return;
-        }
-
-        let thrain_id = gm.pm.current_player.id_name.clone();
-        // Disable dodge & critical variance for determinism
-        gm.pm.current_player.stats.all_stats[DODGE].current = 0;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(boss) = gm
-            .pm
-            .active_bosses
-            .iter_mut()
-            .find(|b| !b.stats.is_dead().unwrap_or(false))
-        {
-            boss.character_rounds_info.is_current_target = true;
-        }
-
-        // Turn 1: Thraïn attacks with "Frappe Cinglante " (trailing space matches filename)
-        let ra1 = gm.launch_attack(Some("Frappe Cinglante "));
-        let aggro_t1 = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name == thrain_id)
-            .map(|h| h.stats.all_stats[AGGRO].current)
-            .unwrap_or(0);
-        assert!(
-            !ra1.new_game_atk_effects.is_empty(),
-            "Frappe Cinglante should produce at least one effect"
-        );
-
-        // Advance through all rounds until Thraïn can attack again (next turn).
-        let mut max_rounds = 60;
-        while gm.pm.current_player.id_name != thrain_id
-            && gm.game_state.status != GameStatus::EndOfGame
-            && gm.game_state.status != GameStatus::EndOfScenario
-            && max_rounds > 0
-        {
-            gm.launch_attack(None);
-            max_rounds -= 1;
-        }
-        if gm.game_state.status == GameStatus::EndOfGame
-            || gm.game_state.status == GameStatus::EndOfScenario
-            || gm.pm.current_player.id_name != thrain_id
-        {
-            return; // game ended, skip
-        }
-
-        gm.pm.current_player.stats.all_stats[DODGE].current = 0;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(boss) = gm
-            .pm
-            .active_bosses
-            .iter_mut()
-            .find(|b| !b.stats.is_dead().unwrap_or(false))
-        {
-            boss.character_rounds_info.is_current_target = true;
-        }
-
-        // Turn 2: Thraïn attacks again
-        let _ra2 = gm.launch_attack(Some("Frappe Cinglante "));
-        let aggro_t2 = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name == thrain_id)
-            .map(|h| h.stats.all_stats[AGGRO].current)
-            .unwrap_or(0);
-
-        assert!(
-            aggro_t2 >= aggro_t1,
-            "Thraïn aggro must not decrease between turns: t1={aggro_t1}, t2={aggro_t2}"
-        );
-        assert!(
-            aggro_t2 > 0,
-            "Thraïn aggro must be > 0 after two attacks: {aggro_t2}"
-        );
-    }
-
-    // ── Rameau Guérisseur tests ────────────────────────────────────────────────
-
-    #[test]
-    fn unit_rameau_guerisseur_initial_heal_range() {
-        let (mut gm, hero_id, _) = testing_test_ally1_vs_test_boss1();
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(buf) = gm
-            .pm
-            .current_player
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::NextHealAtkIsCrit)
-        {
-            buf.is_passive_enabled = false;
-        }
-        gm.pm.set_targeted_characters(&hero_id, "Rameau Guérisseur");
-        let old_hp = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-
-        let ra = gm.launch_attack(Some("Rameau Guérisseur"));
-
-        // Two effects: HP HOT + Magic power buff
-        assert_eq!(ra.new_game_atk_effects.len(), 2, "expected 2 effects");
-
-        let hot = ra
-            .new_game_atk_effects
-            .iter()
-            .find(|g| {
-                g.processed_effect_param
-                    .input_effect_param
-                    .buffer
-                    .stats_name
-                    == HP
-            })
-            .expect("HP effect missing");
-        let per_tick = hot.effect_outcome.full_amount_tx;
-
-        // applies ∈ [1, 3], per apply = 10  → per_tick ∈ [10, 30]
-        assert!(
-            per_tick >= 10,
-            "per-tick heal below minimum (1 apply × 10): {per_tick}"
-        );
-        assert!(
-            per_tick <= 30,
-            "per-tick heal above maximum (3 applies × 10): {per_tick}"
-        );
-
-        // HP was immediately increased on launch
-        let new_hp = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let hp_max = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-        assert_eq!(
-            new_hp,
-            (old_hp as i64 + per_tick).clamp(0, hp_max as i64) as u64
-        );
-    }
-
-    #[test]
-    fn unit_rameau_guerisseur_magic_power_buff() {
-        // The ChangeMaxStat effect shares the ApplyEffectInit count
-        // set by DecreasingRateOnTurn, so full_amount = applies * 10.
-        // With old_magic_max = 30 (20 raw + 10 equipment):
-        //   increase = 30 * (applies * 10) / 100 = applies * 3  → new ∈ [33, 39]
-        let (mut gm, hero_id, _) = testing_test_ally1_vs_test_boss1();
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(buf) = gm
-            .pm
-            .current_player
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::NextHealAtkIsCrit)
-        {
-            buf.is_passive_enabled = false;
-        }
-
-        let old_magic_max = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_POWER]
-            .max;
-
-        gm.pm.set_targeted_characters(&hero_id, "Rameau Guérisseur");
-        let ra = gm.launch_attack(Some("Rameau Guérisseur"));
-
-        // Derive number_of_applies from the HOT's per-tick amount:
-        // per_tick = applies * 10  →  applies = per_tick / 10
-        let per_tick = ra
-            .new_game_atk_effects
-            .iter()
-            .find(|g| {
-                g.processed_effect_param
-                    .input_effect_param
-                    .buffer
-                    .stats_name
-                    == HP
-            })
-            .unwrap()
-            .effect_outcome
-            .full_amount_tx;
-        let applies = per_tick / 10;
-
-        let new_magic_max = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_POWER]
-            .max;
-
-        // full_amount for magic buf = applies * 10 → increase = old * full_amount / 100
-        let full_amount = applies * 10;
-        let expected = old_magic_max + old_magic_max * full_amount as u64 / 100;
-        assert_eq!(
-            new_magic_max,
-            expected,
-            "Magic power should increase by {}% (applies={applies}): {old_magic_max} → {expected}, got {new_magic_max}",
-            applies * 10
-        );
-        // Sanity: increase is proportional to applies (1-3)
-        assert!(
-            new_magic_max >= old_magic_max + old_magic_max * 10 / 100,
-            "min expected +10% increase: got {new_magic_max}"
-        );
-        assert!(
-            new_magic_max <= old_magic_max + old_magic_max * 30 / 100,
-            "max expected +30% increase: got {new_magic_max}"
-        );
-    }
-
-    #[test]
-    fn unit_rameau_guerisseur_hot_lasts_exactly_4_turns() {
-        // The effect entry persists for exactly nb_turns=4 turns regardless of how many
-        // ticks actually fired (which is probabilistic: 1–3). Expiry is always at T5.
-        let (mut gm, hero_id, _) = testing_test_ally1_vs_test_boss1();
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(buf) = gm
-            .pm
-            .current_player
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::NextHealAtkIsCrit)
-        {
-            buf.is_passive_enabled = false;
-        }
-        gm.pm.set_targeted_characters(&hero_id, "Rameau Guérisseur");
-        gm.launch_attack(Some("Rameau Guérisseur"));
-
-        // Advance to test2's first round in T1 — HOT skipped (same launch turn)
-        while gm.pm.current_player.id_name != "test2_#1" {
-            gm.new_round();
-        }
-        assert_eq!(
-            gm.pm.current_player.character_rounds_info.all_effects.len(),
-            2,
-            "both effects must be present in T1"
-        );
-
-        // T2, T3, T4: effects still active when test2 plays each turn
-        for turn_idx in 2..=4 {
-            gm.start_new_turn();
-            while gm.pm.current_player.id_name != "test2_#1" {
-                gm.new_round();
-            }
-            assert_eq!(
-                gm.pm.current_player.character_rounds_info.all_effects.len(),
-                2,
-                "both effects must still be active at turn {turn_idx}"
-            );
-        }
-
-        // T5: counter reaches nb_turns=4 → both effects removed before HOT fires
-        gm.start_new_turn();
-        while gm.pm.current_player.id_name != "test2_#1" {
-            gm.new_round();
-        }
-        assert!(
-            gm.pm
-                .current_player
-                .character_rounds_info
-                .all_effects
-                .is_empty(),
-            "effects must expire after exactly 4 turns (nb_turns=4)"
-        );
-    }
-
-    #[test]
-    fn unit_rameau_guerisseur_hot_fires_at_most_3_ticks() {
-        // Verifies the HOT fires AT MOST 3 times (T2, T3, T4) but not necessarily
-        // exactly 3: the DecreasingRateOnTurn probability means T2=100%, T3=67%,
-        // T4=33%. So the HOT fires 1–3 times depending on the random rolls.
-        let (mut gm, hero_id, _) = testing_test_ally1_vs_test_boss1();
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        if let Some(buf) = gm
-            .pm
-            .current_player
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::NextHealAtkIsCrit)
-        {
-            buf.is_passive_enabled = false;
-        }
-        gm.pm.set_targeted_characters(&hero_id, "Rameau Guérisseur");
-        gm.launch_attack(Some("Rameau Guérisseur"));
-
-        // Skip T1 (HOT does not fire same turn as launch)
-        while gm.pm.current_player.id_name != "test2_#1" {
-            gm.new_round();
-        }
-
-        // HP regen per turn for test2 (7); HOT tick ≥10 — any increase > regen means HOT fired
-        let regen = gm
-            .pm
-            .get_active_hero_character("test2_#1")
-            .unwrap()
-            .stats
-            .all_stats[HP_REGEN]
-            .current as i64;
-
-        let mut hot_ticks = 0u32;
-        for _ in 2..=4 {
-            // Capture HP before start_new_turn because test2_#1 is first in new order:
-            // start_new_turn processes round=1 (test2_#1) which applies HOT immediately.
-            let hp_before = gm
-                .pm
-                .get_active_hero_character("test2_#1")
-                .unwrap()
-                .stats
-                .all_stats[HP]
-                .current as i64;
-            gm.start_new_turn();
-            // After start_new_turn, test2_#1 is current (round=1) with HOT+regen applied.
-            let hp_after = gm
-                .pm
-                .get_active_hero_character("test2_#1")
-                .unwrap()
-                .stats
-                .all_stats[HP]
-                .current as i64;
-            // HOT (≥10) + regen (7) >> regen alone (7)
-            if hp_after - hp_before > regen {
-                hot_ticks += 1;
-            }
-        }
-
-        assert!(
-            hot_ticks >= 1,
-            "HOT must fire at least once (T2 is always 100%): fired {hot_ticks} times"
-        );
-        assert!(
-            hot_ticks <= 3,
-            "HOT must fire at most 3 times (T2–T4): fired {hot_ticks} times"
-        );
-    }
-
-    /// Offrande vitale must boost Thraïn's magical and physical armor max by +50%.
-    #[test]
-    fn unit_offrande_vitale_buffs_thrain_armor() {
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance to Elara's turn
-        let mut max_setup = 30;
-        while !gm.pm.current_player.id_name.contains("Elara") && max_setup > 0 {
-            gm.new_round();
-            max_setup -= 1;
-        }
-        if !gm.pm.current_player.id_name.contains("Elara") {
-            return;
-        }
-
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        gm.pm.current_player.stats.all_stats[MANA].current = mana_max;
-
-        let thrain_id = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name.contains("Thraïn"))
-            .map(|h| h.id_name.clone())
-            .expect("Thraïn must be present");
-
-        let old_mag_armor = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_ARMOR]
-            .max;
-        let old_phy_armor = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[PHYSICAL_ARMOR]
-            .max;
-
-        // Offrande vitale targets a single ally — manually mark Thraïn as the current target.
-        gm.pm
-            .get_mut_active_hero_character(&thrain_id)
-            .unwrap()
-            .character_rounds_info
-            .is_current_target = true;
-
-        gm.launch_attack(Some("Offrande vitale"));
-
-        let new_mag_armor = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_ARMOR]
-            .max;
-        let new_phy_armor = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[PHYSICAL_ARMOR]
-            .max;
-
-        assert_eq!(
-            old_mag_armor + old_mag_armor * 50 / 100,
-            new_mag_armor,
-            "Offrande vitale must boost Thraïn magic armor max by 50%"
-        );
-        assert_eq!(
-            old_phy_armor + old_phy_armor * 50 / 100,
-            new_phy_armor,
-            "Offrande vitale must boost Thraïn physical armor max by 50%"
-        );
-    }
-
-    /// Bouclier Défensif must give exactly +40 aggro to Thraïn (not +42 from implicit Berserk aggro).
-    #[test]
-    fn unit_bouclier_defensif_exact_aggro() {
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance to Thraïn's turn
-        let mut max_setup = 30;
-        while !gm.pm.current_player.id_name.contains("Thraïn") && max_setup > 0 {
-            gm.new_round();
-            max_setup -= 1;
-        }
-        if !gm.pm.current_player.id_name.contains("Thraïn") {
-            return;
-        }
-
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        let thrain_id = gm.pm.current_player.id_name.clone();
-        let aggro_before = gm.pm.current_player.stats.all_stats[AGGRO].current;
-
-        // Bouclier Défensif targets Self — no explicit target setting needed.
-        gm.launch_attack(Some("Bouclier Défensif "));
-
-        let aggro_after = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[AGGRO]
-            .current;
-
-        assert_eq!(
-            aggro_before + 40,
-            aggro_after,
-            "Bouclier Défensif must give exactly +40 aggro (not inflated by Berserk implicit aggro)"
-        );
-    }
-
-    /// Fureur Déchaînée targets Self: no enemy is harmed, Thraïn's Physical power
-    /// max increases by 30 %, and his aggro increases by the explicit +5 aggro effect.
-    #[test]
-    fn unit_fureur_dechainee_self_only() {
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance to Thraïn's turn (hard limit to avoid an infinite loop).
-        let mut max_rounds = 30;
-        while !gm.pm.current_player.id_name.contains("Thraïn") && max_rounds > 0 {
-            gm.new_round();
-            max_rounds -= 1;
-        }
-        if !gm.pm.current_player.id_name.contains("Thraïn") {
-            return;
-        }
-
-        // No crit so the result is deterministic.
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        // Ensure enough Berserk for the attack (cost = 12).
-        gm.pm.current_player.stats.all_stats[BERSERK].current = 50;
-
-        let thrain_id = gm.pm.current_player.id_name.clone();
-
-        let old_phy_pow_max = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[PHYSICAL_POWER]
-            .max;
-        let old_aggro = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[AGGRO]
-            .current;
-
-        // Record every boss HP before the attack.
-        let boss_hp_before: Vec<(String, u64)> = gm
-            .pm
-            .active_bosses
-            .iter()
-            .map(|b| (b.id_name.clone(), b.stats.all_stats[HP].current))
-            .collect();
-
-        // Launch Fureur Déchaînée (target: Self — attack name has two trailing spaces).
-        let result = gm.launch_attack(Some("Fureur Déchaînée  "));
-
-        // --- No effect must land on any enemy ---
-        for gae in &result.new_game_atk_effects {
-            let target = &gae.effect_outcome.target_id_name;
-            assert!(
-                gm.pm.get_active_boss_character(target).is_none(),
-                "Fureur Déchaînée must not affect any boss; got effect on '{target}'"
-            );
-        }
-
-        // Boss HP must be unchanged.
-        for (boss_id, hp_before) in &boss_hp_before {
-            let hp_after = gm
-                .pm
-                .get_active_boss_character(boss_id)
-                .map(|b| b.stats.all_stats[HP].current)
-                .unwrap_or(*hp_before);
-            assert_eq!(
-                *hp_before, hp_after,
-                "Boss '{boss_id}' HP must be unchanged after Fureur Déchaînée"
-            );
-        }
-
-        // --- Self-buff: Physical power max must be +30 % ---
-        let new_phy_pow_max = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[PHYSICAL_POWER]
-            .max;
-        assert_eq!(
-            old_phy_pow_max + old_phy_pow_max * 30 / 100,
-            new_phy_pow_max,
-            "Fureur Déchaînée must boost Thraïn's Physical power max by 30 %"
-        );
-
-        // --- Explicit aggro effect: +5 aggro on Thraïn ---
-        let new_aggro = gm
-            .pm
-            .get_active_hero_character(&thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[AGGRO]
-            .current;
-        assert_eq!(
-            old_aggro + 5,
-            new_aggro,
-            "Fureur Déchaînée must give Thraïn exactly +5 aggro"
-        );
-    }
-
-    // -------------------------------------------------------------------------
-    // Eveil de la forêt (Thalia) — integration tests
-    // -------------------------------------------------------------------------
-
-    fn setup_thalia_turn() -> (super::GameManager, String) {
-        use crate::testing::testing_all_characters::dxrpg_game_manager;
-
-        let mut gm = dxrpg_game_manager();
-        gm.start_game();
-
-        let mut max_rounds = 30;
-        while !gm.pm.current_player.id_name.contains("Thalia") && max_rounds > 0 {
-            gm.new_round();
-            max_rounds -= 1;
-        }
-        // If Thalia never became current player the test is a no-op (guard at call site).
-        let thalia_id = gm.pm.current_player.id_name.clone();
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        gm.pm.current_player.stats.all_stats[MANA].current = mana_max;
-        (gm, thalia_id)
-    }
-
-    /// Eveil de la forêt boosts Magic power max by +10% on every ally.
-    #[test]
-    fn unit_eveil_foret_boosts_magic_power_all_allies() {
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        let old_mag_pow: Vec<(String, u64)> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .filter(|h| h.id_name != thalia_id)
-            .map(|h| (h.id_name.clone(), h.stats.all_stats[MAGICAL_POWER].max))
-            .collect();
-        let old_thalia_mag_pow = gm.pm.current_player.stats.all_stats[MAGICAL_POWER].max;
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        for (id, old_val) in &old_mag_pow {
-            let new_val =
-                gm.pm.get_active_hero_character(id).unwrap().stats.all_stats[MAGICAL_POWER].max;
-            assert_eq!(
-                old_val + old_val * 10 / 100,
-                new_val,
-                "Eveil de la forêt must boost {id} Magic power max by 10%"
-            );
-        }
-        // Also applies to the caster herself (All allies target includes self)
-        let new_thalia_mag_pow = gm
-            .pm
-            .get_active_hero_character(&thalia_id)
-            .unwrap()
-            .stats
-            .all_stats[MAGICAL_POWER]
-            .max;
-        assert_eq!(
-            old_thalia_mag_pow + old_thalia_mag_pow * 10 / 100,
-            new_thalia_mag_pow,
-            "Eveil de la forêt must boost Thalia's own Magic power max by 10%"
-        );
-    }
-
-    /// Eveil de la forêt applies a +80 HP HOT (4 turns) to every ally except the caster.
-    /// The "Ally" + Zone target kind intentionally excludes the launcher.
-    #[test]
-    fn unit_eveil_foret_hot_on_all_allies() {
-        use crate::character_mod::effect::is_hot;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        for hero in gm
-            .pm
-            .active_heroes
-            .iter()
-            .filter(|h| h.id_name != thalia_id)
-        {
-            let hp_hot_count = hero
-                .character_rounds_info
-                .all_effects
-                .iter()
-                .filter(|gae| {
-                    is_hot(
-                        &gae.processed_effect_param.input_effect_param.buffer.kind,
-                        &gae.processed_effect_param
-                            .input_effect_param
-                            .buffer
-                            .stats_name,
-                        gae.processed_effect_param.input_effect_param.buffer.value,
-                    )
-                })
-                .count();
-            assert!(
-                hp_hot_count >= 1,
-                "Eveil de la forêt must apply at least one HP HOT on {}",
-                hero.id_name
-            );
-
-            let hot_effect = hero
-                .character_rounds_info
-                .all_effects
-                .iter()
-                .find(|gae| {
-                    is_hot(
-                        &gae.processed_effect_param.input_effect_param.buffer.kind,
-                        &gae.processed_effect_param
-                            .input_effect_param
-                            .buffer
-                            .stats_name,
-                        gae.processed_effect_param.input_effect_param.buffer.value,
-                    )
-                })
-                .unwrap();
-            assert_eq!(
-                hot_effect
-                    .processed_effect_param
-                    .input_effect_param
-                    .nb_turns,
-                4,
-                "HOT from Eveil de la forêt must last 4 turns on {}",
-                hero.id_name
-            );
-        }
-    }
-
-    /// Eveil de la forêt removes one debuff from every ally except the caster.
-    /// RemoveOneDebuf uses "Ally" + Zone which intentionally excludes the launcher.
-    #[test]
-    fn unit_eveil_foret_removes_one_debuff_from_all_allies() {
-        use crate::character_mod::buffers::BufKinds;
-        use crate::character_mod::effect::is_debuf_effect;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        // Inject a DOT debuff on each ally except Thalia (she is excluded by "Ally" zone).
-        for hero in gm.pm.active_heroes.iter_mut() {
-            if hero.id_name.contains("Thalia") {
-                continue;
-            }
-            let dot = crate::character_mod::effect::ProcessedEffectParam {
-                input_effect_param: crate::character_mod::effect::EffectParam {
-                    nb_turns: 3,
-                    buffer: crate::character_mod::buffers::Buffer {
-                        kind: BufKinds::ChangeCurrentStat,
-                        value: -20,
-                        is_percent: false,
-                        stats_name: HP.to_owned(),
-                        is_passive_enabled: false,
-                        is_passive: false,
-                    },
-                    ..Default::default()
-                },
-                number_of_applies: 1,
-                ..Default::default()
-            };
-            let gae = crate::server::players_manager::GameAtkEffect {
-                processed_effect_param: dot,
-                atk_type: Default::default(),
-                launching_turn: 1,
-                launching_round: 1,
-                effect_outcome: Default::default(),
-            };
-            hero.character_rounds_info.all_effects.push(gae);
-        }
-
-        // Verify each ally now has a debuff.
-        for hero in gm
-            .pm
-            .active_heroes
-            .iter()
-            .filter(|h| !h.id_name.contains("Thalia"))
-        {
-            assert!(
-                hero.character_rounds_info
-                    .all_effects
-                    .iter()
-                    .any(|gae| is_debuf_effect(&gae.processed_effect_param.input_effect_param)),
-                "Setup: {} must have a debuff before launch",
-                hero.id_name
-            );
-        }
-
-        let debuff_counts_before: Vec<(String, usize)> = gm
-            .pm
-            .active_heroes
-            .iter()
-            .filter(|h| !h.id_name.contains("Thalia"))
-            .map(|h| {
-                let count = h
-                    .character_rounds_info
-                    .all_effects
-                    .iter()
-                    .filter(|gae| is_debuf_effect(&gae.processed_effect_param.input_effect_param))
-                    .count();
-                (h.id_name.clone(), count)
-            })
-            .collect();
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        for (id, count_before) in &debuff_counts_before {
-            let count_after = gm
-                .pm
-                .get_active_hero_character(id)
-                .unwrap()
-                .character_rounds_info
-                .all_effects
-                .iter()
-                .filter(|gae| is_debuf_effect(&gae.processed_effect_param.input_effect_param))
-                .count();
-            assert_eq!(
-                count_before - 1,
-                count_after,
-                "Eveil de la forêt must remove exactly one debuff from {id}"
-            );
-        }
-    }
-
-    /// Eveil de la forêt puts a 10-turn cooldown on Thalia.
-    #[test]
-    fn unit_eveil_foret_sets_cooldown_on_thalia() {
-        use crate::character_mod::buffers::BufKinds;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        let thalia = gm.pm.get_active_hero_character(&thalia_id).unwrap();
-        let cooldown_effect = thalia.character_rounds_info.all_effects.iter().find(|gae| {
-            gae.processed_effect_param.input_effect_param.buffer.kind
-                == BufKinds::CooldownTurnsNumber
-        });
-        assert!(
-            cooldown_effect.is_some(),
-            "Eveil de la forêt must apply a CooldownTurnsNumber effect on Thalia"
-        );
-        assert_eq!(
-            cooldown_effect
-                .unwrap()
-                .processed_effect_param
-                .input_effect_param
-                .nb_turns,
-            10,
-            "Eveil de la forêt cooldown must last 10 turns"
-        );
-    }
-
-    /// Eveil de la forêt boosts all active HOTs by +33% via BoostHotsByPercentage.
-    /// Only allies other than the caster are checked because the HOT is "Ally" zone
-    /// (excludes self); "All allies" zone effects do include the caster.
-    #[test]
-    fn unit_eveil_foret_boosts_hots_by_33_percent() {
-        use crate::character_mod::buffers::BufKinds;
-        use crate::character_mod::effect::is_hot;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        // Pre-seed each ally (including current_player so modify_active_character
-        // does not overwrite Thalia's entry) with a HOT so there is something to boost.
-        let pre_hot_ep = crate::character_mod::effect::ProcessedEffectParam {
-            input_effect_param: crate::character_mod::effect::EffectParam {
-                nb_turns: 4,
-                buffer: crate::character_mod::buffers::Buffer {
-                    kind: BufKinds::ChangeCurrentStat,
-                    value: 60,
-                    is_percent: false,
-                    stats_name: HP.to_owned(),
-                    is_passive_enabled: false,
-                    is_passive: false,
-                },
-                ..Default::default()
-            },
-            number_of_applies: 1,
-            ..Default::default()
-        };
-        let pre_hot_gae = crate::server::players_manager::GameAtkEffect {
-            processed_effect_param: pre_hot_ep,
-            atk_type: Default::default(),
-            launching_turn: 1,
-            launching_round: 1,
-            effect_outcome: Default::default(),
-        };
-        for hero in gm.pm.active_heroes.iter_mut() {
-            hero.character_rounds_info
-                .all_effects
-                .push(pre_hot_gae.clone());
-        }
-        gm.pm
-            .current_player
-            .character_rounds_info
-            .all_effects
-            .push(pre_hot_gae);
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        // After launch, every non-caster hero's pre-seeded HOT (60) should be boosted to 79
-        // (60 + 33% of 60 = 79 floor). The BoostHotsByPercentage effect has "All allies"
-        // zone, so it fires for Thalia too — but we only check non-caster heroes here since
-        // the Eveil HOT (value=80) lands only on non-caster allies.
-        for hero in gm
-            .pm
-            .active_heroes
-            .iter()
-            .filter(|h| h.id_name != thalia_id)
-        {
-            let boosted_hot = hero
-                .character_rounds_info
-                .all_effects
-                .iter()
-                .filter(|gae| {
-                    is_hot(
-                        &gae.processed_effect_param.input_effect_param.buffer.kind,
-                        &gae.processed_effect_param
-                            .input_effect_param
-                            .buffer
-                            .stats_name,
-                        gae.processed_effect_param.input_effect_param.buffer.value,
-                    ) && gae.processed_effect_param.input_effect_param.buffer.value >= 79
-                        && gae.processed_effect_param.input_effect_param.buffer.value < 100
-                })
-                .count();
-            assert!(
-                boosted_hot >= 1,
-                "Eveil de la forêt must have boosted the pre-seeded HOT (60→79) on {}",
-                hero.id_name
-            );
-        }
-    }
-
-    /// Eveil de la forêt reinitialises existing HP HOT counters (ReinitBuf effect).
-    /// ReinitBuf uses "All allies" zone so it applies to every ally including the caster.
-    #[test]
-    fn unit_eveil_foret_reinit_hot_counters() {
-        use crate::character_mod::buffers::BufKinds;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        // Seed a HOT that is partially consumed (counter_turn = 2).
-        // Must be added to both active_heroes AND current_player so that
-        // modify_active_character (which copies current_player → active_heroes[thalia])
-        // does not overwrite Thalia's entry and erase the seed.
-        let aged_hot_ep = crate::character_mod::effect::ProcessedEffectParam {
-            input_effect_param: crate::character_mod::effect::EffectParam {
-                nb_turns: 4,
-                buffer: crate::character_mod::buffers::Buffer {
-                    kind: BufKinds::ChangeCurrentStat,
-                    value: 50,
-                    is_percent: false,
-                    stats_name: HP.to_owned(),
-                    is_passive_enabled: false,
-                    is_passive: false,
-                },
-                ..Default::default()
-            },
-            counter_turn: 2,
-            number_of_applies: 1,
-            ..Default::default()
-        };
-        let aged_hot_gae = crate::server::players_manager::GameAtkEffect {
-            processed_effect_param: aged_hot_ep,
-            atk_type: Default::default(),
-            launching_turn: 1,
-            launching_round: 1,
-            effect_outcome: Default::default(),
-        };
-        for hero in gm.pm.active_heroes.iter_mut() {
-            hero.character_rounds_info
-                .all_effects
-                .push(aged_hot_gae.clone());
-        }
-        gm.pm
-            .current_player
-            .character_rounds_info
-            .all_effects
-            .push(aged_hot_gae);
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        // ReinitBuf resets counter_turn to 0 on every HP HOT for all allies.
-        for hero in &gm.pm.active_heroes {
-            let reset = hero.character_rounds_info.all_effects.iter().any(|gae| {
-                gae.processed_effect_param.input_effect_param.buffer.kind
-                    == BufKinds::ChangeCurrentStat
-                    && gae
-                        .processed_effect_param
-                        .input_effect_param
-                        .buffer
-                        .stats_name
-                        == HP
-                    && gae.processed_effect_param.input_effect_param.buffer.value > 0
-                    && gae.processed_effect_param.counter_turn == 0
-            });
-            assert!(
-                reset,
-                "Eveil de la forêt ReinitBuf must reset counter_turn=0 on HP HOT for {}",
-                hero.id_name
-            );
-        }
-    }
-
-    /// Eveil de la forêt sets a BoostedByHots buffer on Thalia proportional to her active HOT count.
-    #[test]
-    fn unit_eveil_foret_boosted_by_hots_on_thalia() {
-        use crate::character_mod::buffers::BufKinds;
-        use crate::character_mod::effect::is_hot;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        // Count HOTs currently active on Thalia before launch.
-        let hot_count_before = gm
-            .pm
-            .current_player
-            .character_rounds_info
-            .all_effects
-            .iter()
-            .filter(|gae| {
-                is_hot(
-                    &gae.processed_effect_param.input_effect_param.buffer.kind,
-                    &gae.processed_effect_param
-                        .input_effect_param
-                        .buffer
-                        .stats_name,
-                    gae.processed_effect_param.input_effect_param.buffer.value,
-                )
-            })
-            .count() as i64;
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        // BoostBufByHotsNumberInPercentage fires BEFORE the zone HOT is added to Thalia,
-        // so BoostedByHots value = hot_count_before * 20.
-        let thalia = gm.pm.get_active_hero_character(&thalia_id).unwrap();
-        let boosted = thalia
-            .character_rounds_info
-            .get_buffer_by_type(&BufKinds::BoostedByHots);
-        assert!(
-            boosted.is_some(),
-            "Eveil de la forêt must set a BoostedByHots buffer on Thalia"
-        );
-        assert_eq!(
-            hot_count_before * 20,
-            boosted.unwrap().value,
-            "BoostedByHots value must equal number_of_hots × 20"
-        );
-    }
-
-    /// Eveil de la forêt (BoostHotsByPercentage +33%) must boost a pre-existing HOT on
-    /// Azrak Ombresang — simulating the HOT he would have received from Essence Régénératrice.
-    /// This covers the zone-target bug where only the caster's HOTs were previously boosted.
-    #[test]
-    fn unit_eveil_foret_boosts_azrak_existing_hot() {
-        use crate::character_mod::buffers::BufKinds;
-        use crate::character_mod::effect::is_hot;
-
-        let (mut gm, thalia_id) = setup_thalia_turn();
-        if !thalia_id.contains("Thalia") {
-            return;
-        }
-
-        let azrak_id = gm
-            .pm
-            .active_heroes
-            .iter()
-            .find(|h| h.id_name.contains("Azrak"))
-            .map(|h| h.id_name.clone())
-            .expect("Azrak Ombresang must be in the lotr party");
-
-        // Seed a +12 HP HOT on Azrak (as Essence Régénératrice would give).
-        let hot_value: i64 = 12;
-        let hot_ep = crate::character_mod::effect::ProcessedEffectParam {
-            input_effect_param: crate::character_mod::effect::EffectParam {
-                nb_turns: 4,
-                buffer: crate::character_mod::buffers::Buffer {
-                    kind: BufKinds::ChangeCurrentStat,
-                    value: hot_value,
-                    is_percent: false,
-                    stats_name: HP.to_owned(),
-                    is_passive_enabled: false,
-                    is_passive: false,
-                },
-                ..Default::default()
-            },
-            number_of_applies: 1,
-            ..Default::default()
-        };
-        let hot_gae = crate::server::players_manager::GameAtkEffect {
-            processed_effect_param: hot_ep,
-            atk_type: Default::default(),
-            launching_turn: 1,
-            launching_round: 1,
-            effect_outcome: crate::character_mod::effect::EffectOutcome {
-                full_amount_tx: hot_value,
-                real_amount_tx: hot_value,
-                target_id_name: azrak_id.clone(),
-                ..Default::default()
-            },
-        };
-        gm.pm
-            .get_mut_active_hero_character(&azrak_id)
-            .unwrap()
-            .character_rounds_info
-            .all_effects
-            .push(hot_gae);
-
-        gm.launch_attack(Some("Eveil de la forêt"));
-
-        // After Eveil de la forêt, Azrak's HOT value must be boosted by +33%.
-        let azrak = gm.pm.get_active_hero_character(&azrak_id).unwrap();
-        let azrak_hot = azrak
-            .character_rounds_info
-            .all_effects
-            .iter()
-            .find(|gae| {
-                is_hot(
-                    &gae.processed_effect_param.input_effect_param.buffer.kind,
-                    &gae.processed_effect_param
-                        .input_effect_param
-                        .buffer
-                        .stats_name,
-                    gae.processed_effect_param.input_effect_param.buffer.value,
-                ) && gae.processed_effect_param.input_effect_param.buffer.value
-                    >= hot_value + hot_value * 33 / 100
-                    && gae.processed_effect_param.input_effect_param.buffer.value
-                        <= hot_value + hot_value * 33 / 100 + 1
-            })
-            .expect("Azrak's HOT must be boosted by +33% by Eveil de la forêt");
-
-        let boosted = hot_value + hot_value * 33 / 100;
-        assert_eq!(
-            boosted,
-            azrak_hot
-                .processed_effect_param
-                .input_effect_param
-                .buffer
-                .value,
-            "Azrak's HOT buffer.value must be boosted from {hot_value} to {boosted} (+33%)"
-        );
-        assert_eq!(
-            boosted, azrak_hot.effect_outcome.full_amount_tx,
-            "Azrak's HOT effect_outcome.full_amount_tx must be boosted so ticks heal {boosted} HP"
-        );
-        assert_eq!(
-            boosted, azrak_hot.effect_outcome.real_amount_tx,
-            "Azrak's HOT effect_outcome.real_amount_tx must be boosted so log_text() shows {boosted} HP"
-        );
-    }
-
-    /// Récupération Mordorienne: for each enemy killed last turn, Azrak gains
-    /// +10% Physical power (2 turns) and recovers 10% HP, Vigor, and Mana instantly.
-    /// This test uses 2 kills, so all bonuses are doubled.
-    #[test]
-    fn unit_azrak_recuperation_mordorienne_2_kills() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.start_game();
-
-        // Advance until Azrak is current player.
-        let mut max_rounds = 30;
-        while !gm.pm.current_player.id_name.contains("Azrak") && max_rounds > 0 {
-            gm.new_round();
-            max_rounds -= 1;
-        }
-        if !gm.pm.current_player.id_name.contains("Azrak") {
-            return;
-        }
-        let azrak_id = gm.pm.current_player.id_name.clone();
-
-        // Simulate 2 enemy kills on the previous turn.
-        let prev_turn = gm.game_state.current_turn_nb.saturating_sub(1);
-        gm.game_state
-            .died_ennemies
-            .insert(prev_turn, vec!["Boss_A".to_owned(), "Boss_B".to_owned()]);
-
-        // Snapshot max stats and Physical power base (raw + equip, no buf_effect) before the attack.
-        let hp_max = gm.pm.current_player.stats.all_stats[HP].max;
-        let vigor_max = gm.pm.current_player.stats.all_stats[VIGOR].max;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        let phy_pow_max_before = gm.pm.current_player.stats.all_stats[PHYSICAL_POWER].max;
-        // base_value = raw + equip (without existing buf_effect_percent, which the % is applied on top of)
-        let phy_pow_attr = &gm.pm.current_player.stats.all_stats[PHYSICAL_POWER];
-        let phy_base = phy_pow_attr.max_raw as i64
-            + phy_pow_attr.buf_equip_value
-            + phy_pow_attr.buf_equip_percent * phy_pow_attr.max_raw as i64 / 100;
-
-        // Drain resources to ~50% so recovery is clearly visible and stays below cap.
-        gm.pm.current_player.stats.all_stats[HP].current = hp_max / 2;
-        gm.pm.current_player.stats.all_stats[VIGOR].current = vigor_max / 2;
-        gm.pm.current_player.stats.all_stats[MANA].current = mana_max / 2;
-        let hp_before = gm.pm.current_player.stats.all_stats[HP].current;
-        let vigor_before = gm.pm.current_player.stats.all_stats[VIGOR].current;
-        let mana_before = gm.pm.current_player.stats.all_stats[MANA].current;
-
-        // Disable crits for determinism.
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        gm.launch_attack(Some("Récupération Mordorienne"));
-
-        let azrak = gm.pm.get_active_hero_character(&azrak_id).unwrap();
-
-        // 2 kills × 10% = +20% Physical power max for 2 turns.
-        // ChangeMaxStat applies to base_value (raw + equip), not current max.
-        let phy_pow_max_after = azrak.stats.all_stats[PHYSICAL_POWER].max;
-        let expected_pow_max = phy_pow_max_before + (phy_base * 20 / 100) as u64;
-        assert_eq!(
-            expected_pow_max, phy_pow_max_after,
-            "Physical power max must rise by 20% of base (2 kills × 10%)"
-        );
-
-        // 2 kills × 10% = instant +20% of max for each resource.
-        // ChangeCurrentStat: full_amount = max * value / 100
-        let expected_hp_gain = hp_max * 20 / 100;
-        assert_eq!(
-            hp_before + expected_hp_gain,
-            azrak.stats.all_stats[HP].current,
-            "HP must recover by 20% of max HP"
-        );
-
-        let expected_vigor_gain = vigor_max * 20 / 100;
-        assert_eq!(
-            vigor_before + expected_vigor_gain,
-            azrak.stats.all_stats[VIGOR].current,
-            "Vigor must recover by 20% of max Vigor"
-        );
-
-        let expected_mana_gain = mana_max * 20 / 100;
-        assert_eq!(
-            mana_before + expected_mana_gain,
-            azrak.stats.all_stats[MANA].current,
-            "Mana must recover by 20% of max Mana"
-        );
-    }
-
-    /// Integration test: 3 heroes (Elara, Azrak, Thalia) + 1 enemy.
-    /// Elara's IsDamageTxHealNeedyAlly passive fires immediately when she deals damage,
-    /// healing the most-needy alive ally and emitting a log in the same turn.
-    #[test]
-    fn unit_passive_damage_tx_heal_needy_3_heroes_1_enemy() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(1);
-
-        let elara_id = "Elara_la_guerisseuse_de_la_Lorien_#1";
-        let azrak_id = "Azrak_Ombresang_#1";
-
-        // Drain Azrak to 10 HP — lowest ratio → most needy
-        gm.pm
-            .get_mut_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats
-            .get_mut(HP)
-            .unwrap()
-            .current = 10;
-        let azrak_hp_max = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-
-        // Set Elara as current player, no crit
-        let elara = gm.pm.get_active_hero_character(elara_id).unwrap().clone();
-        gm.pm.current_player = elara;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        // Boss: no dodge, no armor bonus, no DamageRxPercent
-        let boss_id = gm.pm.active_bosses[0].id_name.clone();
-        gm.pm.active_bosses[0].stats.all_stats[DODGE].current = 0;
-        gm.pm.active_bosses[0].stats.all_stats[MAGICAL_ARMOR].current = 0;
-        gm.pm.active_bosses[0]
-            .character_rounds_info
-            .is_current_target = true;
-        if let Some(buf) = gm.pm.active_bosses[0]
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::DamageRxPercent)
-        {
-            buf.value = 0;
-        }
-
-        let old_boss_hp = gm.pm.active_bosses[0].stats.all_stats[HP].current;
-
-        let result = gm.launch_attack(Some("Frappe élémentaire"));
-
-        // Derive actual damage from boss HP delta (avoids hardcoding hero magic power)
-        let new_boss_hp = gm
-            .pm
-            .get_active_boss_character(&boss_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let actual_damage = old_boss_hp as i64 - new_boss_hp as i64;
-        assert!(actual_damage > 0, "SimpleAtk must deal damage to the boss");
-
-        // Passive fires in the same turn: 25% of damage heals Azrak (most needy)
-        let expected_heal = (actual_damage as u64 * 25 / 100).min(azrak_hp_max - 10);
-        let new_azrak_hp = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        assert_eq!(
-            new_azrak_hp,
-            10 + expected_heal,
-            "Elara's passive must heal Azrak by 25% of damage dealt in the same turn"
-        );
-
-        // Passive log must appear in logs_atk of this attack result
-        assert!(
-            result
-                .logs_atk
-                .iter()
-                .any(|l| l.message.contains("Passive")),
-            "passive heal log must appear in logs_atk"
-        );
-    }
-
-    /// Integration test: Thraïn's passive ChangeCurrentStat(Dodge, 10) raises his
-    /// effective Dodge from 5 to 15 at load time and survives an equipment toggle.
-    #[test]
-    fn unit_passive_dodge_stat_thrain_3_heroes_1_enemy() {
-        use crate::testing::testing_all_characters::testing_all_equipment;
-
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Thraïn_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(1);
-
-        let thrain_id = "Thraïn_#1";
-
-        // At load: base_value = max_raw(5) + equip(24: amulet+4, cape+10, shoes+10) = 29
-        // passive: +10% of 29 = 2 (integer) → total = 31
-        let dodge_after_load = gm
-            .pm
-            .get_active_hero_character(thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[DODGE]
-            .current;
-        assert_eq!(31, dodge_after_load, "passive must be included at load");
-
-        // After removing the starting amulet (Dodge +4), base_value drops to 25.
-        // passive: +10% of 25 = 2 (integer) → total = 27
-        let thrain = gm.pm.get_mut_active_hero_character(thrain_id).unwrap();
-        thrain.toggle_equipment("starting amulet", &testing_all_equipment());
-        let dodge_after_toggle = thrain.stats.all_stats[DODGE].current;
-        assert_eq!(
-            27, dodge_after_toggle,
-            "Dodge must be 27 after removing amulet (passive still applies)"
-        );
-    }
-
-    #[test]
-    fn unit_elara_frappe_elementaire_3_heroes_3_enemies() {
-        use crate::character_mod::rounds_information::AmountType;
-
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let elara_id = "Elara_la_guerisseuse_de_la_Lorien_#1";
-        let elara = gm.pm.get_active_hero_character(elara_id).unwrap().clone();
-        gm.pm.current_player = elara;
-        gm.pm.current_player.level = 100;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        // Set target: first boss, no dodge
-        let boss_id = gm.pm.active_bosses[0].id_name.clone();
-        gm.pm.active_bosses[0].stats.all_stats[DODGE].current = 0;
-        gm.pm.active_bosses[0]
-            .character_rounds_info
-            .is_current_target = true;
-        // Angmar carries a DamageRxPercent:100 passive that would double damage;
-        // zero it out so this test verifies the base armor-mitigation formula only.
-        if let Some(buf) = gm.pm.active_bosses[0]
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::DamageRxPercent)
-        {
-            buf.value = 0;
-        }
-
-        let old_boss_hp = gm.pm.active_bosses[0].stats.all_stats[HP].current;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        let old_mana = gm.pm.current_player.stats.all_stats[MANA].current;
-
-        // Expected magic damage: rebalanced base value=76, Elara mag power, boss mag armor+pow
-        let hero_mag_pow = gm.pm.current_player.stats.get_power_stat(true);
-        let boss_mag_armor = gm.pm.active_bosses[0].stats.get_armor_stat(true);
-        let boss_mag_pow = gm.pm.active_bosses[0].stats.get_power_stat(true);
-        let power_factor = 1.0 + hero_mag_pow as f64 / AttackType::POWER_SCALE;
-        let raw_dmg = (76_f64 * power_factor).round() as i64;
-        let defense = boss_mag_armor as f64 + boss_mag_pow as f64 / AttackType::DEFENSE_DIVISOR;
-        let protection = AttackType::ARMOR_FACTOR / (AttackType::ARMOR_FACTOR + defense);
-        let expected_dmg = (raw_dmg as f64 * protection).round() as i64;
-
-        // RepeatIfHeal does not fire: no heal on turn 0
-        gm.pm.current_player.character_rounds_info.tx_rx[AmountType::HealTx as usize].clear();
-
-        let ra = gm.launch_attack(Some("Frappe élémentaire"));
-        assert!(
-            !ra.new_game_atk_effects.is_empty(),
-            "attack must produce effects"
-        );
-
-        // Mana cost: 15% of mana_max
-        assert_eq!(
-            old_mana - 15 * mana_max / 100,
-            gm.pm
-                .get_active_hero_character(elara_id)
-                .unwrap()
-                .stats
-                .all_stats[MANA]
-                .current,
-            "Elara mana cost: 15% of max"
-        );
-
-        // Boss HP reduced by expected magic damage
-        let new_boss_hp = gm
-            .pm
-            .get_active_boss_character(&boss_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        assert_eq!(
-            old_boss_hp as i64 - expected_dmg,
-            new_boss_hp as i64,
-            "boss HP reduced by magic damage"
-        );
-    }
-
-    #[test]
-    fn unit_elara_don_de_vie_3_heroes_3_enemies() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let elara_id = "Elara_la_guerisseuse_de_la_Lorien_#1";
-        let azrak_id = "Azrak_Ombresang_#1";
-
-        let elara = gm.pm.get_active_hero_character(elara_id).unwrap().clone();
-        gm.pm.current_player = elara;
-        gm.pm.current_player.level = 100;
-
-        // Drain Azrak to 50 HP and mark as ally target
-        gm.pm
-            .get_mut_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current = 50;
-        gm.pm
-            .get_mut_active_hero_character(azrak_id)
-            .unwrap()
-            .character_rounds_info
-            .is_current_target = true;
-
-        let azrak_hp_max = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-        let old_elara_hp = gm.pm.current_player.stats.all_stats[HP].current;
-        let elara_hp_max = gm.pm.current_player.stats.all_stats[HP].max;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        let old_mana = gm.pm.current_player.stats.all_stats[MANA].current;
-
-        gm.launch_attack(Some("Don de vie"));
-
-        let new_azrak_hp = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let new_elara_hp = gm
-            .pm
-            .get_active_hero_character(elara_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let new_mana = gm
-            .pm
-            .get_active_hero_character(elara_id)
-            .unwrap()
-            .stats
-            .all_stats[MANA]
-            .current;
-
-        // DecreasingRateOnTurn(3): minimum 1 apply (first roll is always 100%)
-        let min_heal_one_apply = 30 * azrak_hp_max / 100;
-        assert!(new_azrak_hp > 50, "Azrak should be healed above 50 HP");
-        assert!(new_azrak_hp <= azrak_hp_max, "Azrak HP must not exceed max");
-        assert!(
-            new_azrak_hp >= 50 + min_heal_one_apply,
-            "Azrak must receive at least 1 apply (30% HP max = {min_heal_one_apply})"
-        );
-
-        // Elara takes -15% HP max self-damage (at least 1 apply)
-        let min_self_dmg = 15 * elara_hp_max / 100;
-        assert!(
-            new_elara_hp < old_elara_hp,
-            "Elara should take self-damage from Don de vie"
-        );
-        assert!(
-            old_elara_hp - new_elara_hp >= min_self_dmg,
-            "Elara self-damage must be at least 1 apply (15% HP max = {min_self_dmg})"
-        );
-
-        // Mana cost: 24% of mana_max
-        assert_eq!(
-            old_mana - 24 * mana_max / 100,
-            new_mana,
-            "Elara mana cost: 24% of max"
-        );
-    }
-
-    #[test]
-    fn unit_elara_lumiere_curative_3_heroes_3_enemies() {
-        use crate::character_mod::rounds_information::AmountType;
-
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let elara_id = "Elara_la_guerisseuse_de_la_Lorien_#1";
-        let azrak_id = "Azrak_Ombresang_#1";
-
-        let elara = gm.pm.get_active_hero_character(elara_id).unwrap().clone();
-        gm.pm.current_player = elara;
-        gm.pm.current_player.level = 100;
-
-        // Sub-case A: condition not met (turn 0, no DamageTx on previous turn)
-        let lumiere_atk = gm
-            .pm
-            .current_player
-            .attacks_list
-            .get("Lumiere curative")
-            .unwrap()
-            .clone();
-        assert!(
-            !gm.pm.current_player.can_be_launched(&lumiere_atk, 0),
-            "Lumiere curative must be blocked when no prior damage (turn 0)"
-        );
-        assert!(
-            !gm.pm.current_player.can_be_launched(&lumiere_atk, 1),
-            "Lumiere curative must be blocked when DamageTx[0] is absent"
-        );
-
-        // Sub-case B: condition met — inject DamageTx[1] = 200 and set turn 2
-        gm.pm.current_player.character_rounds_info.tx_rx[AmountType::DamageTx as usize]
-            .insert(1, 200);
-        gm.game_state.current_turn_nb = 2;
-        assert!(
-            gm.pm.current_player.can_be_launched(&lumiere_atk, 2),
-            "Lumiere curative must be launchable when DamageTx[prev_turn] > 0"
-        );
-
-        // Drain Azrak to 10 HP and mark as ally target
-        gm.pm
-            .get_mut_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current = 10;
-        gm.pm
-            .get_mut_active_hero_character(azrak_id)
-            .unwrap()
-            .character_rounds_info
-            .is_current_target = true;
-
-        let azrak_hp_max = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        let old_mana = gm.pm.current_player.stats.all_stats[MANA].current;
-        let mana_regen = gm.pm.current_player.stats.all_stats[MANA_REGEN].current;
-        // Heal formula adds launcher magical power: full_amount = (buffer.value + pow_current)
-        let elara_mag_pow = gm.pm.current_player.stats.get_power_stat(true);
-        let heal_amount = (130 + elara_mag_pow) as u64;
-
-        gm.launch_attack(Some("Lumiere curative"));
-
-        let new_azrak_hp = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let new_mana = gm
-            .pm
-            .get_active_hero_character(elara_id)
-            .unwrap()
-            .stats
-            .all_stats[MANA]
-            .current;
-
-        // Heal = 130 + Elara's magical power (heal formula: buffer.value + pow_current), capped at max
-        let expected_hp = std::cmp::min(10 + heal_amount, azrak_hp_max);
-        assert_eq!(
-            expected_hp, new_azrak_hp,
-            "Azrak must be healed for {} HP (capped at {azrak_hp_max})",
-            heal_amount
-        );
-
-        // Mana cost 15% of max; eval_end_of_round applies regen so account for mana_regen
-        let expected_mana =
-            (old_mana as i64 - (15 * mana_max / 100) as i64 + mana_regen as i64) as u64;
-        assert_eq!(expected_mana, new_mana, "Elara mana: cost 15% + regen");
-    }
-
-    #[test]
-    fn unit_elara_non_sans_raison_3_heroes_3_enemies() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1" | "Thalia_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let elara_id = "Elara_la_guerisseuse_de_la_Lorien_#1";
-        let azrak_id = "Azrak_Ombresang_#1";
-        let thalia_id = "Thalia_#1";
-
-        let elara = gm.pm.get_active_hero_character(elara_id).unwrap().clone();
-        gm.pm.current_player = elara;
-        gm.pm.current_player.level = 100;
-
-        // Drain all heroes to 10 HP (zone heal must restore all to max)
-        gm.pm.current_player.stats.all_stats[HP].current = 10;
-        for h in gm.pm.active_heroes.iter_mut() {
-            h.stats.all_stats[HP].current = 10;
-        }
-
-        let elara_hp_max = gm.pm.current_player.stats.all_stats[HP].max;
-        let azrak_hp_max = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-        let thalia_hp_max = gm
-            .pm
-            .get_active_hero_character(thalia_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .max;
-        let mana_max = gm.pm.current_player.stats.all_stats[MANA].max;
-        let old_mana = gm.pm.current_player.stats.all_stats[MANA].current;
-
-        gm.launch_attack(Some("Non sans raison"));
-
-        // All heroes healed to max HP (100% of HP max heal > remaining 90% deficit)
-        let new_elara_hp = gm
-            .pm
-            .get_active_hero_character(elara_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let new_azrak_hp = gm
-            .pm
-            .get_active_hero_character(azrak_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        let new_thalia_hp = gm
-            .pm
-            .get_active_hero_character(thalia_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        assert_eq!(elara_hp_max, new_elara_hp, "Elara must be healed to max HP");
-        assert_eq!(azrak_hp_max, new_azrak_hp, "Azrak must be healed to max HP");
-        assert_eq!(
-            thalia_hp_max, new_thalia_hp,
-            "Thalia must be healed to max HP"
-        );
-
-        // Elara has BlockHealAtk active (heals blocked for 3 turns)
-        assert!(
-            gm.pm
-                .get_active_hero_character(elara_id)
-                .unwrap()
-                .character_rounds_info
-                .is_heal_atk_blocked,
-            "Elara heal attacks must be blocked after Non sans raison"
-        );
-
-        // Mana cost: 24% of max mana
-        assert_eq!(
-            old_mana - 24 * mana_max / 100,
-            gm.pm
-                .get_active_hero_character(elara_id)
-                .unwrap()
-                .stats
-                .all_stats[MANA]
-                .current,
-            "Non sans raison costs 24% of max mana"
-        );
-    }
-
-    #[test]
-    fn unit_thrain_enchainement_furieux_3_heroes_3_enemies() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Thraïn_#1" | "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let thrain_id = "Thraïn_#1";
-        let thrain = gm.pm.get_active_hero_character(thrain_id).unwrap().clone();
-        gm.pm.current_player = thrain;
-        gm.pm.current_player.level = 100;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        // Set berserk to 60; the attack costs 20% of max berserk per apply and fires as many
-        // times as possible (RepeatAsManyAsPossible) until rage is exhausted.
-        // actual_cost = raw_cost * berserk_max / 100; nb_applies = floor(60 / actual_cost).
-        gm.pm.current_player.stats.all_stats[BERSERK].current = 60;
-        let berserk_max = gm.pm.current_player.stats.all_stats[BERSERK].max;
-        let berseck_cost = 20u64;
-        let actual_cost = (berseck_cost * berserk_max / 100).max(1);
-        let nb_applies = (60u64 / actual_cost).max(1);
-
-        // Target: first boss, zero dodge and DamageRxPercent for clean formula
-        let boss_id = gm.pm.active_bosses[0].id_name.clone();
-        let old_boss_hp = gm.pm.active_bosses[0].stats.all_stats[HP].current;
-        gm.pm.active_bosses[0].stats.all_stats[DODGE].current = 0;
-        gm.pm.active_bosses[0]
-            .character_rounds_info
-            .is_current_target = true;
-        if let Some(buf) = gm.pm.active_bosses[0]
-            .character_rounds_info
-            .get_mut_buffer_by_type(&BufKinds::DamageRxPercent)
-        {
-            buf.value = 0;
-        }
-
-        gm.launch_attack(Some("Enchaînement Furieux"));
-
-        // RepeatAsManyAsPossible bypasses armor; each apply deals the raw 50 damage.
-        let expected_dmg = nb_applies * 50;
-        let new_boss_hp = gm
-            .pm
-            .get_active_boss_character(&boss_id)
-            .unwrap()
-            .stats
-            .all_stats[HP]
-            .current;
-        assert_eq!(
-            old_boss_hp - expected_dmg,
-            new_boss_hp,
-            "boss HP should drop by {nb_applies} × 50 = {expected_dmg} (RepeatAsManyAsPossible)"
-        );
-
-        // Every apply drains rage: total cost = nb_applies × actual_cost.
-        let new_berserk = gm
-            .pm
-            .get_active_hero_character(thrain_id)
-            .unwrap()
-            .stats
-            .all_stats[BERSERK]
-            .current;
-        let expected_berserk = 60u64.saturating_sub(nb_applies * actual_cost);
-        assert_eq!(
-            expected_berserk, new_berserk,
-            "Thraïn berserk: {nb_applies} applies × {actual_cost} cost each, expected {expected_berserk}"
-        );
-    }
-
-    #[test]
-    fn unit_thrain_provocation_feroce_3_heroes_3_enemies() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Thraïn_#1" | "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let thrain_id = "Thraïn_#1";
-        let thrain = gm.pm.get_active_hero_character(thrain_id).unwrap().clone();
-        gm.pm.current_player = thrain;
-        gm.pm.current_player.level = 100;
-
-        let old_berserk = gm.pm.current_player.stats.all_stats[BERSERK].current;
-        let old_crit_max = gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].max;
-        let old_aggro = gm.pm.current_player.stats.all_stats[AGGRO].current;
-
-        // Init aggro tx_rx slot so process_aggro can update the stat
-        gm.pm
-            .current_player
-            .init_aggro_on_turn(gm.game_state.current_turn_nb);
-
-        gm.launch_attack(Some("Provocation Féroce "));
-
-        let thrain_after = gm.pm.get_active_hero_character(thrain_id).unwrap();
-
-        // +12 Berserk (free attack, no cost)
-        assert_eq!(
-            old_berserk + 12,
-            thrain_after.stats.all_stats[BERSERK].current,
-            "Thraïn berserk must increase by 12 (no cost)"
-        );
-
-        // +10 Aggro on self
-        assert_eq!(
-            old_aggro + 10,
-            thrain_after.stats.all_stats[AGGRO].current,
-            "Thraïn aggro must increase by 10"
-        );
-
-        // +40 max Critical strike for 3 turns
-        assert_eq!(
-            old_crit_max + 40,
-            thrain_after.stats.all_stats[CRITICAL_STRIKE].max,
-            "Thraïn critical strike max must increase by 40"
-        );
-
-        // 5-turn cooldown applied
-        let cooldown_active = thrain_after
-            .character_rounds_info
-            .all_effects
-            .iter()
-            .any(|e| {
-                e.processed_effect_param.input_effect_param.buffer.kind
-                    == BufKinds::CooldownTurnsNumber
-                    && e.atk_type.name.contains("Provocation")
-            });
-        assert!(
-            cooldown_active,
-            "Provocation Féroce must have a 5-turn cooldown"
-        );
-    }
-
-    #[test]
-    fn unit_thrain_tourbillon_destructeur_3_heroes_3_enemies() {
-        let mut gm = testing_all_characters::dxrpg_game_manager();
-        gm.pm.active_heroes.retain(|h| {
-            matches!(
-                h.id_name.as_str(),
-                "Thraïn_#1" | "Elara_la_guerisseuse_de_la_Lorien_#1" | "Azrak_Ombresang_#1"
-            )
-        });
-        gm.pm.active_bosses.truncate(3);
-
-        let thrain_id = "Thraïn_#1";
-        let thrain = gm.pm.get_active_hero_character(thrain_id).unwrap().clone();
-        gm.pm.current_player = thrain;
-        gm.pm.current_player.level = 100;
-        gm.pm.current_player.stats.all_stats[CRITICAL_STRIKE].current = 0;
-
-        let thrain_phy_pow = gm.pm.current_player.stats.get_power_stat(false);
-        let old_berserk = gm.pm.current_player.stats.all_stats[BERSERK].current;
-        let berserk_max = gm.pm.current_player.stats.all_stats[BERSERK].max;
-        let berseck_cost = 15u64;
-        let cost_deducted = berseck_cost * berserk_max / 100;
-
-        let old_berserk_rate_pct =
-            gm.pm.current_player.stats.all_stats[BERSECK_RATE].buf_effect_percent;
-
-        // Zero out dodge and DamageRxPercent on all 3 bosses for clean damage formula
-        let old_boss_hps: Vec<u64> = gm
-            .pm
-            .active_bosses
-            .iter()
-            .map(|b| b.stats.all_stats[HP].current)
-            .collect();
-        let boss_phy_armors: Vec<i64> = gm
-            .pm
-            .active_bosses
-            .iter()
-            .map(|b| b.stats.get_armor_stat(false))
-            .collect();
-        let boss_phy_pows: Vec<i64> = gm
-            .pm
-            .active_bosses
-            .iter()
-            .map(|b| b.stats.get_power_stat(false))
-            .collect();
-        for boss in gm.pm.active_bosses.iter_mut() {
-            boss.stats.all_stats[DODGE].current = 0;
-            if let Some(buf) = boss
-                .character_rounds_info
-                .get_mut_buffer_by_type(&BufKinds::DamageRxPercent)
-            {
-                buf.value = 0;
-            }
-        }
-
-        // Init aggro slot so the +5 Aggro self-effect can be recorded
-        gm.pm
-            .current_player
-            .init_aggro_on_turn(gm.game_state.current_turn_nb);
-        let old_aggro = gm.pm.current_player.stats.all_stats[AGGRO].current;
-
-        gm.launch_attack(Some("Tourbillon Destructeur "));
-
-        // All 3 bosses take physical damage with rebalanced base value 67
-        for (i, ((old_hp, phy_armor), boss_phy_pow)) in old_boss_hps
-            .iter()
-            .zip(boss_phy_armors.iter())
-            .zip(boss_phy_pows.iter())
-            .enumerate()
-        {
-            let boss_id = gm.pm.active_bosses[i].id_name.clone();
-            let new_hp = gm
-                .pm
-                .get_active_boss_character(&boss_id)
-                .unwrap()
-                .stats
-                .all_stats[HP]
-                .current;
-            let power_factor = 1.0 + thrain_phy_pow as f64 / AttackType::POWER_SCALE;
-            let raw_dmg = (67_f64 * power_factor).round() as i64;
-            let defense = *phy_armor as f64 + *boss_phy_pow as f64 / AttackType::DEFENSE_DIVISOR;
-            let protection = AttackType::ARMOR_FACTOR / (AttackType::ARMOR_FACTOR + defense);
-            let expected_dmg = (raw_dmg as f64 * protection).round() as i64;
-            // HP is floored at 0 when damage exceeds current HP
-            let expected_hp = (*old_hp as i64 - expected_dmg).max(0);
-            assert_eq!(
-                expected_hp, new_hp as i64,
-                "boss[{i}] HP must drop by {expected_dmg} (physical zone damage)"
-            );
-        }
-
-        let thrain_after = gm.pm.get_active_hero_character(thrain_id).unwrap();
-
-        // Berserk cost: 15% of max
-        assert_eq!(
-            old_berserk - cost_deducted,
-            thrain_after.stats.all_stats[BERSERK].current,
-            "Thraïn berserk: 15% of max deducted"
-        );
-
-        // +5 explicit Aggro on self; zone damage also generates implicit aggro
-        assert!(
-            thrain_after.stats.all_stats[AGGRO].current >= old_aggro + 5,
-            "Thraïn aggro must increase by at least 5 (explicit self effect)"
-        );
-
-        // +100% max Berserk rate for 4 turns: buf_effect_percent increased by 100
-        assert_eq!(
-            old_berserk_rate_pct + 100,
-            thrain_after.stats.all_stats[BERSECK_RATE].buf_effect_percent,
-            "Berserk rate buf_effect_percent must increase by 100"
         );
     }
 }
